requirements.txt
requests
python-dotenv
termcolor
Pillow
google-genai
duckduckgo-search
transformers
torch
